新アーキテクチャへのリファクタ #55
