+++ app/login.tsx
@@ -1,0 +1,5 @@
+import { LoginScreen } from '@/src/features/auth';
+
+export default function LoginPage() {
+  return <LoginScreen />;
+}
