+++ src/api/index.ts
@@ -1,0 +1,2 @@
+export { apiClient } from './apiClient';
+export * from './generated/apiSchema';
