+++ src/features/auth/app-logo/index.ts
@@ -1,0 +1,1 @@
+export { AppLogo } from './AppLogo';
