+++ src/features/auth/feature-list/index.ts
@@ -1,0 +1,1 @@
+export { FeatureList } from './FeatureList';
