+++ src/features/auth/google-login-button/index.ts
@@ -1,0 +1,1 @@
+export { GoogleLoginButton } from './GoogleLoginButton';
