+++ src/features/home/folder-list/index.ts
@@ -1,0 +1,1 @@
+export { FolderList } from './FolderList';
