+++ src/features/home/recent-notes/index.ts
@@ -1,0 +1,1 @@
+export { RecentNotes } from './RecentNotes';
