+++ src/features/home/record-fab/index.ts
@@ -1,0 +1,1 @@
+export { RecordFab } from './RecordFab';
