+++ src/features/home/sample-data/index.ts
@@ -1,0 +1,2 @@
+export { sampleNotes, sampleUser, folderStructure } from './sampleData';
+export type { FolderItem, CategoryItem } from './sampleData';
