+++ src/features/home/user-header/index.ts
@@ -1,0 +1,1 @@
+export { UserHeader } from './UserHeader';
