+++ src/features/notes/index.ts
@@ -1,0 +1,1 @@
+export { NoteDetailScreen } from './NoteDetailScreen';
