+++ src/features/notes/note-content/index.ts
@@ -1,0 +1,1 @@
+export { NoteContent } from './NoteContent';
