+++ src/features/notes/note-header/index.ts
@@ -1,0 +1,1 @@
+export { NoteHeader } from './NoteHeader';
