+++ src/features/notes/note-tags/NoteTags.tsx
@@ -1,0 +1,70 @@
+import { colors } from '@/src/shared/constants';
+import { useState } from 'react';
+import { TextInput as RNTextInput, View } from 'react-native';
+import { Chip } from 'react-native-paper';
+
+// 型定義
+interface NoteTagsProps {
+  tags: string[];
+  onAddTag: (tag: string) => void;
+  onRemoveTag: (tag: string) => void;
+}
+
+// コンポーネント
+export function NoteTags({ tags, onAddTag, onRemoveTag }: NoteTagsProps) {
+  const [isAddingTag, setIsAddingTag] = useState(false);
+  const [newTag, setNewTag] = useState('');
+
+  const handleAddTag = () => {
+    if (newTag.trim()) {
+      onAddTag(newTag.trim());
+      setNewTag('');
+      setIsAddingTag(false);
+    }
+  };
+
+  return (
+    <View className="mb-4">
+      <View className="flex-row flex-wrap gap-2">
+        {tags.map((tag) => (
+          <Chip
+            key={tag}
+            onClose={() => onRemoveTag(tag)}
+            className="bg-t-bg-tertiary"
+            textStyle={{ color: colors.text.primary, fontSize: 14 }}
+          >
+            #{tag}
+          </Chip>
+        ))}
+        {isAddingTag ? (
+          <View className="mr-2">
+            <RNTextInput
+              value={newTag}
+              onChangeText={setNewTag}
+              onSubmitEditing={handleAddTag}
+              onBlur={() => {
+                if (newTag.trim()) {
+                  handleAddTag();
+                } else {
+                  setIsAddingTag(false);
+                }
+              }}
+              placeholder="タグ名"
+              className="px-3 py-2 border border-t-brand-500 rounded-lg bg-t-bg-primary text-sm min-w-[100px]"
+              autoFocus
+            />
+          </View>
+        ) : (
+          <Chip
+            icon="plus"
+            onPress={() => setIsAddingTag(true)}
+            className="bg-t-bg-primary border border-t-border-primary border-dashed"
+            textStyle={{ color: colors.text.secondary, fontSize: 14 }}
+          >
+            タグを追加
+          </Chip>
+        )}
+      </View>
+    </View>
+  );
+}
