+++ src/features/notes/note-tags/index.ts
@@ -1,0 +1,1 @@
+export { NoteTags } from './NoteTags';
