+++ src/features/notes/sample-data/index.ts
@@ -1,0 +1,2 @@
+export { sampleNote } from './sampleData';
+export type { NoteData } from './sampleData';
