+++ src/features/processing/processing-spinner/index.ts
@@ -1,0 +1,1 @@
+export { ProcessingSpinner } from './ProcessingSpinner';
