+++ src/features/recording/audio-recorder/index.ts
@@ -1,0 +1,2 @@
+export { audioRecorderService } from './AudioRecorderService';
+export type { RecordingResult, AudioLevel } from './AudioRecorderService';
