+++ src/features/recording/audio-waveform/index.ts
@@ -1,0 +1,1 @@
+export { AudioWaveform } from './AudioWaveform';
