+++ src/features/recording/index.ts
@@ -1,0 +1,3 @@
+export { RecordingScreen } from './RecordingScreen';
+export { audioRecorderService } from './audio-recorder';
+export type { RecordingResult, AudioLevel } from './audio-recorder';
