+++ src/features/recording/recording-controls/index.ts
@@ -1,0 +1,1 @@
+export { RecordingControls } from './RecordingControls';
