+++ src/features/recording/recording-timer/index.ts
@@ -1,0 +1,1 @@
+export { RecordingTimer } from './RecordingTimer';
