+++ src/shared/constants/index.ts
@@ -1,0 +1,2 @@
+export { colors } from './colors';
+export type { Colors } from './colors';
